with cecyn1@hotmail.com SVN user with carlos@ca2.cc Microsoft Account at "east"-us OVH Windows 8 VMware ESXi guest!! Eu te amo carlinhos (carlinhos é sempre Carlos Gustavo Cecyn Lundgren para mim)!! Inspired by the lite of Carlos Gustavo Cecyn Lundgren!! by Camilo!! Inpiration make me create default folder ca2 at default drive Z for building always ca2 and having access to executables, by now, like appfy (app._.exe)!!
Eu te amo Carlinhos!!

Eu te amo Mummi!!

Obrigado Rafael Cecyn Lundgren!!

+++ appseed/base/base/crypto/crypto_crypto.cpp
@@ -652,7 +652,7 @@ namespace crypto
 
          ctx.Init();
 
-         ctx.update(psz, strlen(psz));
+         ctx.update(psz,(int32_t) strlen(psz));
 
          memSha1.set(0, memSha1.get_size());
