+++ appseed/base/base/openssl/engines/ccgost/gosthash.c
@@ -188,7 +188,7 @@ int hash_block(gost_hash_ctx *ctx,const byte *block, size_t length)
 		unsigned int add_bytes = 32-ctx->left;
 		if (add_bytes>length)
 			{
-			add_bytes = length;
+			add_bytes = (unsigned int) length;
 			}
 		memcpy(&(ctx->remainder[ctx->left]),block,add_bytes);
 		ctx->left+=add_bytes;
@@ -212,7 +212,7 @@ int hash_block(gost_hash_ctx *ctx,const byte *block, size_t length)
 		}
 	if (curptr!=block+length)
 		{
-		ctx->left=block+length-curptr;
+		ctx->left= (int) (block+length-curptr);
 		memcpy(ctx->remainder,curptr,ctx->left);
 		}
 	return 1;
